fix: ssr of home events for runtime-env deployments
