+++ src/app/[locale]/(platform)/_lib/dynamic-home-category-page.tsx
@@ -1,12 +1,7 @@
 import type { Metadata } from 'next'
 import type { SupportedLocale } from '@/i18n/locales'
-import { cacheLife } from 'next/cache'
 import { notFound } from 'next/navigation'
-import HomeContent from '@/app/[locale]/(platform)/(home)/_components/HomeContent'
-import {
-  getHomeInitialCurrentTimestamp,
-  HOME_INITIAL_EVENTS_CACHE_LIFE,
-} from '@/app/[locale]/(platform)/(home)/_utils/homeInitialEventsCache'
+import HomeInitialContent from '@/app/[locale]/(platform)/(home)/_components/HomeInitialContent'
 import {
   buildLocalizedPagePath,
   buildPredictionResultsOgImageUrl,
@@ -23,29 +18,6 @@ import { getPublicShellStaticParams, shouldBypassPublicShellPlaceholder, STATIC_
 async function getMainTags(locale: SupportedLocale) {
   const { data: mainTags } = await loadPlatformMainTags(locale)
   return mainTags ?? []
-}
-
-async function CachedDynamicHomeContent({
-  initialMainTag,
-  initialTag,
-  locale,
-}: {
-  initialMainTag?: string
-  initialTag: string
-  locale: SupportedLocale
-}) {
-  'use cache'
-  cacheLife(HOME_INITIAL_EVENTS_CACHE_LIFE)
-
-  const currentTimestamp = getHomeInitialCurrentTimestamp()
-  return (
-    <HomeContent
-      locale={locale}
-      currentTimestamp={currentTimestamp}
-      initialTag={initialTag}
-      initialMainTag={initialMainTag}
-    />
-  )
 }
 
 export async function generateDynamicHomeCategoryStaticParams() {
@@ -163,7 +135,7 @@ export async function DynamicHomeCategoryPageContent({
     notFound()
   }
 
-  return <CachedDynamicHomeContent locale={locale} initialTag={category.slug} />
+  return <HomeInitialContent locale={locale} initialTag={category.slug} />
 }
 
 export async function DynamicHomeSubcategoryPageContent({
@@ -193,7 +165,7 @@ export async function DynamicHomeSubcategoryPageContent({
   }
 
   return (
-    <CachedDynamicHomeContent
+    <HomeInitialContent
       locale={locale}
       initialTag={resolvedSubcategory.subcategory.slug}
       initialMainTag={resolvedSubcategory.category.slug}
